frontend/requirements.txt
requests==2.26.0
python-dotenv==0.10.1
